requirements.txt
async-timeout==3.0.1
discord.py==1.7.2
SQLAlchemy==1.4.15
